+++ backend/IntexAPI/IntexAPI/Models/ApiDtos.cs
@@ -4,7 +4,7 @@ namespace IntexAPI.Models;
 public record PublicImpactSummaryDto(
     int ActiveResidentsCount,
     int SafehouseCount,
-    double TotalDonationsAllTime,
+    decimal TotalDonationsAllTime,
     IReadOnlyList<PublicImpactSnapshotSummaryDto> LatestPublishedSnapshots);
 
 public record PublicImpactSnapshotSummaryDto(
@@ -27,7 +27,7 @@ public record RecentDonationDto(
     int DonationId,
     DateOnly? DonationDate,
     string? DonationType,
-    double? Amount,
+    decimal? Amount,
     string? CurrencyCode,
     int? SupporterId,
     string? SupporterDisplayName);
@@ -45,12 +45,12 @@ public record LatestSafehouseProgressDto(
     string? SafehouseName,
     DateOnly? MonthStart,
     int? ActiveResidents,
-    double? AvgEducationProgress,
-    double? AvgHealthScore,
+    decimal? AvgEducationProgress,
+    decimal? AvgHealthScore,
     int? IncidentCount);
 
 // --- Analytics ---
-public record DonationTrendPointDto(int Year, int Month, double TotalAmount, int DonationCount);
+public record DonationTrendPointDto(int Year, int Month, decimal TotalAmount, int DonationCount);
 
 public record ReintegrationStatusCountDto(string? Status, int Count);
 
@@ -59,14 +59,14 @@ public record SafehousePerformanceDto(
     string? SafehouseName,
     DateOnly? MonthStart,
     int? ActiveResidents,
-    double? AvgEducationProgress,
-    double? AvgHealthScore,
+    decimal? AvgEducationProgress,
+    decimal? AvgHealthScore,
     int? ProcessRecordingCount,
     int? HomeVisitationCount,
     int? IncidentCount);
 
 public record OutcomeTrendPointDto(
     DateOnly MonthStart,
-    double? AvgEducationProgress,
-    double? AvgHealthScore,
+    decimal? AvgEducationProgress,
+    decimal? AvgHealthScore,
     int ActiveResidentsSum);
